Use Vector2i only in cross_corr_consistency_check instantiations
Both BM and SGM/MGM call cross_corr_consistency_check with integer
disparity. Float subpixel disparity is produced later. Change the
unit test from Vector2f to Vector2i to match production usage and
drop the Vector2f explicit instantiation. Add explanatory comment.

+++ src/vw/Stereo/tests/TestCorrelate.cxx
@@ -24,18 +24,18 @@
 using namespace vw;
 using namespace vw::stereo;
 
-typedef PixelMask<Vector2f> PixelDisp;
+typedef PixelMask<Vector2i> PixelDisp;
 
 TEST( Correlate, CrossCorrConsistency ) {
 
   ImageView<PixelDisp> r2l(3,3), l2r(3,3);
-  fill( r2l, PixelDisp(Vector2f(0,0)) );
-  fill( l2r, PixelDisp(Vector2f(0,0)) );
+  fill( r2l, PixelDisp(Vector2i(0,0)) );
+  fill( l2r, PixelDisp(Vector2i(0,0)) );
 
-  fill( crop(l2r,2,0,1,3), PixelDisp(Vector2f(2,2)) );
-  l2r(0,0) = PixelDisp(Vector2f(1,1));
-  r2l(1,1) = PixelDisp(Vector2f(-1,-1));
-  l2r(1,0) = PixelDisp(Vector2f(1,1));
+  fill( crop(l2r,2,0,1,3), PixelDisp(Vector2i(2,2)) );
+  l2r(0,0) = PixelDisp(Vector2i(1,1));
+  r2l(1,1) = PixelDisp(Vector2i(-1,-1));
+  l2r(1,0) = PixelDisp(Vector2i(1,1));
 
   ImageView<PixelDisp> l2r_copy = copy(l2r);
   cross_corr_consistency_check( l2r_copy, r2l, 0 );
